LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-cookies-expected.txt

Blocked access to external URL http://www.localhost:8800/fetch/api/resources/inspect-headers.py?cors&headers=cookie

Harness Error (TIMEOUT), message = null

PASS Omit mode: no cookie sent 
FAIL Include mode: 1 cookie assert_equals: Request includes cookie(s) expected (string) "a=1" but got (object) null
PASS Include mode: local cookies are not sent with remote request 
PASS Include mode: remote cookies are not sent with local request 
PASS Same-origin mode: cookies are discarded in cors request 
TIMEOUT Include mode: remote cookies are not sent with other remote request Test timed out
